+++ backend/src/main/java/com/optimiza/clickbarber/model/Barbeiro.java
@@ -1,6 +1,5 @@
 package com.optimiza.clickbarber.model;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import lombok.*;
@@ -18,7 +17,7 @@ public class Barbeiro {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String cpf;
     private String nome;
