+++ backend/src/main/java/com/optimiza/clickbarber/model/dto/cliente/ClienteAgendamentoDto.java
@@ -2,8 +2,6 @@ package com.optimiza.clickbarber.model.dto.cliente;
 
 import lombok.*;
 
-import java.util.UUID;
-
 @Getter
 @Setter
 @ToString
@@ -13,7 +11,7 @@ import java.util.UUID;
 @Builder
 public class ClienteAgendamentoDto {
 
-    private Integer id;
+    private Long id;
     private String cpf;
     private String nome;
     private String email;
